feat(web): make the zoom map the Knowledge Graph at /knowledge-graph
The zoom map takes over as the Knowledge Graph and keeps that route, so every
existing link, breadcrumb and the /c4 redirect stay valid:

- /repos/[id]/knowledge-graph now renders the zoom map (was the old node-link /
  layered-C4 view).
- /repos/[id]/zoom becomes a redirect to /knowledge-graph (forwarding ?focus=),
  so old links keep working.
- Drop the NEXT_PUBLIC_ENABLE_ZOOM_MAP flag and the separate "Zoom Map" nav item;
  the single "Knowledge Graph" nav item points at the route.
- Delete the now-dead web host files for the old view (c4-view.tsx,
  use-architecture.ts, arch-detail-panel-host.tsx). They were the only web
  consumers of @repowise-dev/ui/c4.

The reusable C4 machinery is deliberately untouched: the backend c4_builder, the
zoom map's own /zoom-map endpoint in routers/c4.py, and @repowise-dev/ui/c4 (still
used by the VS Code architecture webview) all stay. /architecture is unaffected.
